refactor: Initial commit of Pasture Collector 1.7+ port
