+++ fabric/build.gradle.kts
@@ -1,0 +1,50 @@
+import org.jetbrains.kotlin.gradle.dsl.JvmTarget
+
+plugins {
+    id("dev.architectury.loom")
+    id("architectury-plugin")
+    id("com.gradleup.shadow") version("9.2.2")
+}
+
+repositories {
+    mavenLocal()
+    mavenCentral()
+    maven("https://dl.cloudsmith.io/public/geckolib3/geckolib/maven/")
+    maven("https://maven.impactdev.net/repository/development/")
+    maven("https://api.modrinth.com/maven")
+}
+
+dependencies {
+    minecraft("com.mojang:minecraft:${property("minecraft_version")}")
+    mappings(loom.officialMojangMappings())
+    modImplementation("net.fabricmc:fabric-loader:${property("fabric_loader_version")}")
+
+    // Fabric
+    modImplementation("net.fabricmc.fabric-api:fabric-api:${property("fabric_api_version")}")
+    modImplementation("net.fabricmc:fabric-language-kotlin:${property("fabric_kotlin_version")}")
+
+    // Mod deps
+    modImplementation("com.cobblemon:fabric:${property("cobblemon_version")}")
+    modImplementation("maven.modrinth:cobblemon-tim-core:${property("tim_core_fabric_version")}")
+    modImplementation("maven.modrinth:cobblemon-droploottables:${property("droploottables_fabric_version")}")
+}
+
+tasks {
+    processResources {
+        inputs.property("version", project.version)
+
+        filesMatching("fabric.mod.json") {
+            expand(mutableMapOf("version" to project.version))
+        }
+    }
+
+    jar {
+        from("LICENSE")
+    }
+
+    compileKotlin {
+        compilerOptions {
+            jvmTarget = JvmTarget.JVM_21
+        }
+    }
+}
