+++ fabric/src/main/kotlin/us/timinc/mc/cobblemon/pasturecollector/fabric/config/PastureCollectorConfig.kt
@@ -1,0 +1,7 @@
+package us.timinc.mc.cobblemon.pasturecollector.fabric.config
+
+open class PastureCollectorConfig {
+    val debug: Boolean = false
+    val baseCobblemonLootEnabled: Boolean = true
+    val chanceToDrop: Float = 1F
+}
