+++ fabric/src/main/kotlin/us/timinc/mc/cobblemon/pasturecollector/fabric/droppers/PastureBlockDropper.kt
@@ -1,0 +1,8 @@
+package us.timinc.mc.cobblemon.pasturecollector.fabric.droppers
+
+import us.timinc.mc.cobblemon.droploottables.api.droppers.AbstractFormDropper
+import us.timinc.mc.cobblemon.pasturecollector.fabric.PastureCollectorMod.MOD_ID
+
+object PastureBlockDropper : AbstractFormDropper("collector", MOD_ID) {
+    override fun load() {}
+}
