+++ fabric/src/main/resources/assets/pasturecollector/blockstates/pasture_collector.json
