+++ fabric/src/main/resources/assets/pasturecollector/models/block/pasture_collector.json
