+++ fabric/src/main/resources/assets/pasturecollector/models/item/iron_slab.json
