+++ fabric/src/main/resources/assets/pasturecollector/models/item/pasture_collector.json
