+++ fabric/src/main/resources/assets/pasturecollector/textures/block/pasture_collector_bottom.png
