+++ fabric/src/main/resources/assets/pasturecollector/textures/block/pasture_collector_side.png
