+++ fabric/src/main/resources/assets/pasturecollector/textures/gui/pasture_collector_inventory.png
