+++ fabric/src/main/resources/data/minecraft/tags/block/mineable/pickaxe.json
