+++ fabric/src/main/resources/data/pasturecollector/loot_table/blocks/pasture_collector.json
