+++ fabric/src/main/resources/data/pasturecollector/recipe/pasture_collector.json
