+++ gradle.properties
@@ -1,18 +1,42 @@
-# Done to increase the memory available to gradle.
-org.gradle.jvmargs=-Xmx2G
+org.gradle.jvmargs=-Xmx4G
+org.gradle.parallel=true
+org.gradle.caching=true
+org.gradle.caching.debug=false
+org.gradle.configureondemand=true
 
-# Fabric Properties
-# check these on https://fabricmc.net/develop
-minecraft_version=1.21.1
-loader_version=0.16.9
+modId=cobblemon_pasturecollector
+modCobblemonVersion=1.7.3
+modMyVersion=1.4.0
+modName=Cobblemon Pasture Collector
+modDescription=Collects items from Pokemon in nearby Pasture Blocks in Cobblemon.
+modAuthor=timinc aka Timothy Metcalfe
+
+# TODO: Change me or riot
+modIssues=https://github.com/timinc-cobble/cobblemon-droploottables/issues
+modSource=https://github.com/timinc-cobble/cobblemon-droploottables
+# TODO: end
+
+modDiscord=https://discord.com/invite/WKAR27SdSv
 
 # Mod Properties
-mod_version=1.6-fabric-1.3.0
+mod_version=1.7-fabric-1.4.0
 maven_group=us.timinc.mc.cobblemon
 archives_base_name=pasturecollector
 
-# Dependencies
-fabric_version=0.114.0+1.21.1
-fabric_kotlin_version=1.9.3+kotlin.1.8.20
-cobblemon_version=1.6.1+1.21.1
-droploottables_version=1.6-fabric-1.1.0
+# Fabric project dependencies
+fabric_loader_version=0.17.2
+fabric_api_version=0.116.6+1.21.1
+fabric_kotlin_version=1.13.9+kotlin.2.3.10
+
+# Neoforge project dependencies
+neoforge_version=21.1.182
+kotlin_for_forge_version=5.10.0
+
+# Common dependencies
+minecraft_version=1.21.1
+cobblemon_version=1.7.3+1.21.1
+junit_version=5.1.0
+tim_core_fabric_version=1.7.3-fabric-1.31.0
+tim_core_neoforge_version=1.7.3-neoforge-1.31.0
+droploottables_fabric_version=1.7.3-fabric-1.7.1
+droploottables_neoforge_version=1.7.3-neoforge-1.7.1
