--- src/main/kotlin/us/timinc/mc/cobblemon/pasturecollector/config/PastureCollectorConfig.kt
@@ -1,7 +1,0 @@
-package us.timinc.mc.cobblemon.pasturecollector.config
-
-open class PastureCollectorConfig {
-    val debug: Boolean = false
-    val baseCobblemonLootEnabled: Boolean = true
-    val chanceToDrop: Float = 1F
-}
