--- src/main/kotlin/us/timinc/mc/cobblemon/pasturecollector/droppers/PastureBlockDropper.kt
@@ -1,8 +1,0 @@
-package us.timinc.mc.cobblemon.pasturecollector.droppers
-
-import us.timinc.mc.cobblemon.droploottables.api.droppers.AbstractFormDropper
-import us.timinc.mc.cobblemon.pasturecollector.PastureCollectorMod.MOD_ID
-
-object PastureBlockDropper : AbstractFormDropper("collector", MOD_ID) {
-    override fun load() {}
-}
